+++ R/animate_distr.R
@@ -27,8 +27,10 @@
 #' distr <- get_distr(bf, ts)
 #' anim <- animate_distr(distr, bf,  show_dynamic_mask = TRUE)
 #'
+#' \dontrun{
+#' # Display it
 #' anim
-#'
+#' }
 #' ### Project a distribution
 #'
 #' # Make starting distribution